added .travis.yml

+++ test-output-sparse.sh
@@ -3,10 +3,18 @@
 echo
 echo "expecting error"
 ./output-sparse
+if [ "$?" == "0" ]; then
+	echo "program exited with error code 0, which was unexpected"
+	exit 1
+fi
 
 echo
 echo "expecting error"
 ./output-sparse -vi input.img output.img
+if [ "$?" == "0" ]; then
+	echo "program exited with error code 0, which was unexpected"
+	exit 1
+fi
 
 echo
 dd if=/dev/zero of=input.img bs=4k count=4 2> /dev/null
